# docs/api/console-api.openapi.yaml
openapi: 3.1.0
info:
  title: Batch Platform Console API
  version: 0.1.0
  description: >
    Frontend-facing contract for the batch console.
    This file is the machine-readable source of truth for UI integration,
    code generation, and mock server stubs.
    Whenever the API surface changes, update this file and console-api-protocol.md together.
servers:
  - url: http://localhost:8080
    description: Local console-api
...
security:
  - bearerAuth: []
  - consoleToken: []
paths:
...
  /api/console/instances/{id}/cancel:
    post:
      tags: [Instances]
      summary: Cancel a job instance
      operationId: cancelJobInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/instances/{id}/terminate:
    post:
      tags: [Instances]
      summary: Terminate a job instance
      operationId: terminateJobInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P0: Trigger management ───────────────────────────────────────────────────
...
  /api/console/batch-windows:
    get:
      tags: [BatchWindows]
      summary: List batch windows
      operationId: listBatchWindows
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: windowCode
          in: query
          required: false
          schema:
            type: string
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Batch window list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseObject'
    post:
      tags: [BatchWindows]
      summary: Create batch window
      operationId: createBatchWindow
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Created batch window id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseLong'
  /api/console/batch-windows/{id}:
    put:
      tags: [BatchWindows]
      summary: Update batch window
      operationId: updateBatchWindow
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/batch-windows/{id}/toggle:
    post:
      tags: [BatchWindows]
      summary: Enable or disable batch window
      operationId: toggleBatchWindow
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P1: Calendars + Holidays ─────────────────────────────────────────────────
...
  /api/console/pipeline-definitions:
    get:
      tags: [PipelineDefinitions]
      summary: List pipeline definitions
      operationId: listPipelineDefinitions
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: jobCode
          in: query
          required: false
          schema:
            type: string
        - name: pipelineType
          in: query
          required: false
          schema:
            type: string
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Pipeline definition list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseObject'
    post:
      tags: [PipelineDefinitions]
      summary: Create pipeline definition
      operationId: createPipelineDefinition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Created pipeline definition id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseLong'
  /api/console/pipeline-definitions/{id}:
    get:
      tags: [PipelineDefinitions]
      summary: Get pipeline definition detail with steps
      operationId: getPipelineDefinition
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pipeline definition detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseObject'
    put:
      tags: [PipelineDefinitions]
      summary: Update pipeline definition
      operationId: updatePipelineDefinition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/pipeline-definitions/{id}/toggle:
    post:
      tags: [PipelineDefinitions]
      summary: Enable or disable pipeline definition
      operationId: togglePipelineDefinition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P2: Partition operations ─────────────────────────────────────────────────
  /api/console/instances/partitions/{id}/cancel:
    post:
      tags: [Instances]
      summary: Cancel a job instance partition
      operationId: cancelJobInstancePartition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
  /api/console/instances/partitions/{id}/retry:
    post:
      tags: [Instances]
      summary: Retry a failed job instance partition
      operationId: retryJobInstancePartition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'

  # ─── P2: Workflow run operations ──────────────────────────────────────────────
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    consoleToken:
      type: apiKey
      in: header
      name: X-Console-Token
  parameters:
...
    PageNoQuery:
      name: pageNo
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 20
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  schemas:
    CommonMeta:
      type: object
      properties:
        requestId:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [requestId, traceId, timestamp]
    CommonResponseBase:
      type: object
      properties:
        code:
          type: string
          description: Application-level status code.
        message:
          type: string
          description: Human-readable status message.
        meta:
          $ref: '#/components/schemas/CommonMeta'
      required: [code, message, meta]
...
    CommonResponseString:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: string
    CommonResponseLong:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              type: integer
              format: int64
...
    CommonResponseObject:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              nullable: true
...